Feature: Split Data by ID Since the above function use random slices to split data, running it a couple of times will ensure the ML algo wil see all of the data.
To avoid this, compute a hash of each instance’s identifier and put that instance in the test set if the hash is lower or equal to 20% of the maximum hash value.

This ensures that the test set will remain consistent across multiple runs, even if you refresh the dataset. The new test set will contain 20% of the new instances, but it will not contain any instance that was previously in the training set.